rename result to preAssessment
