+++ schema/schema.json
@@ -5362,6 +5362,14 @@
       ],
       "type": "object"
     },
+    "PreAssessment": {
+      "$id": "#PreAssessment",
+      "description": "The result of the application based on information provided by the applicant, prior to assessment by a planning officer. Results are determined by flags corresponding to responses; each application can have up to one result per flagset",
+      "items": {
+        "$ref": "#/definitions/ResultFlag"
+      },
+      "type": "array"
+    },
     "ProjectType": {
       "$id": "#ProjectType",
       "anyOf": [
@@ -17571,14 +17579,6 @@
       },
       "type": "array"
     },
-    "Result": {
-      "$id": "#Result",
-      "description": "The result of the application. Results are determined by flags corresponding to responses; each application can have up to one result per flagset",
-      "items": {
-        "$ref": "#/definitions/ResultFlag"
-      },
-      "type": "array"
-    },
     "ResultFlag": {
       "$id": "#ResultFlag",
       "anyOf": [
@@ -18827,19 +18827,18 @@
     "metadata": {
       "$ref": "#/definitions/Metadata"
     },
+    "preAssessment": {
+      "$ref": "#/definitions/PreAssessment"
+    },
     "responses": {
       "$ref": "#/definitions/Responses"
-    },
-    "result": {
-      "$ref": "#/definitions/Result"
     }
   },
   "required": [
     "data",
-    "result",
-    "metadata",
     "responses",
-    "files"
+    "files",
+    "metadata"
   ],
   "title": "Digital Planning Application",
   "type": "object"
